client-pipeline/scripts/requirements.txt
python-docx==0.8.11
openai==1.44.0
python-dotenv>=1.0.1
jsonschema>=4.21.1
httpx==0.27.0
